+++ docs/changelog.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 0.2.2 (May 12, 2024)
+
+* Added server.py containing a REST API for CELI as well as WebSockets for continuous updates
+* Added a celi-react-demo which is a prototype React front end for the Wikipedia use case that demonstrates the server.
+* Added the ability to provide asynchronous human feedback on any section at any time.
+* Added new examples - GRE, AP_English_Language, and AP_English_Literature
+
+## 0.2.1 (May 12, 2024)
+
+* Updated the HumanEval use case with additional results 
+
 ## 0.2.0 (May 6, 2024)
 * Changed to 0.2.0 because there is an API change (tools should now inherit from `BaseDocToolImplementations` instead of `ToolImplementations`)
 * Moved writing of the document into the tools.  This allows CELI to be used for tasks that don't involve document